Split norm benchmarks
