+++ tools/test-op.sh
@@ -68,7 +68,7 @@ for item in $CHANGED_FILES; do
 done
 
 # Skip tests if no tests file is found
-if [ ${#TEST_CASES[@]} -eq 0  && ${#PERF_TEST_CASES[@]} -eq 0 ]; then
+if [[ ${#TEST_CASES[@]} -eq 0  && ${#PERF_TEST_CASES[@]} -eq 0 ]]; then
   exit 0
 fi
